+++ backend/src/collections/user-db/user-db.module.ts
@@ -5,6 +5,7 @@ import { generateRandomString } from 'picsur-shared/dist/util/random';
 import { AuthConfigService } from '../../config/early/auth.config.service.js';
 import { EarlyConfigModule } from '../../config/early/early-config.module.js';
 import { EUserBackend } from '../../database/entities/users/user.entity.js';
+import { ImageDBModule } from '../image-db/image-db.module.js';
 import { PreferenceDbModule } from '../preference-db/preference-db.module.js';
 import { RoleDbModule } from '../role-db/role-db.module.js';
 import { UserDbService } from './user-db.service.js';
@@ -14,6 +15,7 @@ import { UserDbService } from './user-db.service.js';
     EarlyConfigModule,
     RoleDbModule,
     PreferenceDbModule,
+    ImageDBModule,
     TypeOrmModule.forFeature([EUserBackend]),
   ],
   providers: [UserDbService],